requirements.txt
Flask==2.3.3
requests==2.32.3
python-dotenv==1.0.1
joblib==1.4.2
scikit-learn==1.6.1
tensorflow